Update Potential Block Examples
