--- examples/PotentialBlocks/Tunnel/joints/circleCavern.csv
@@ -1,15 +1,0 @@
-centreX;centreY;centreZ;length;traceX;traceY;phi_b;phi_r;JRC;JSC;asperity;simgaC;cohesion;tension;;
-0.0000000e+00;0;1.0000000e+00;10;5.53;7000;0;0;0;0;0;0;0;0;
-1.0000000e+00;0;7.2736615e-16;5;10.53;7000;0;0;0;0;0;0;0;0;
-9.6592583e-01;0;-2.5881905e-01;5;0.53;7000;0;0;0;0;0;0;0;0;
-8.6602540e-01;0;-5.0000000e-01;5;0.53;7000;0;0;0;0;0;0;0;0;
-7.0710678e-01;0;-7.0710678e-01;5;0.53;7000;0;0;0;0;0;0;0;0;
-5.0000000e-01;0;-8.6602540e-01;5;0.53;7000;0;0;0;0;0;0;0;0;
-2.5881905e-01;0;-9.6592583e-01;5;0.53;7000;0;0;0;0;0;0;0;0;
-2.3429107e-15;0;-1.0000000e+00;5;0.53;7000;0;0;0;0;0;0;0;0;
--2.5881905e-01;0;-9.6592583e-01;5;0.53;7000;0;0;0;0;0;0;0;0;
--5.0000000e-01;0;-8.6602540e-01;5;0.53;7000;0;0;0;0;0;0;0;0;
--7.0710678e-01;0;-7.0710678e-01;5;0.53;7000;0;0;0;0;0;0;0;0;
--8.6602540e-01;0;-5.0000000e-01;5;0.53;7000;0;0;0;0;0;0;0;0;
--9.6592583e-01;0;-2.5881905e-01;5;0.53;7000;0;0;0;0;0;0;0;0;
--1.0000000e+00;0;-3.7364107e-15;5;10.53;7000;0;0;0;0;0;0;0;0;
