--- examples/PotentialBlocks/Tunnel/joints/circleOriginal.csv
@@ -1,13 +1,0 @@
-centreX;centreY;centreZ;length;traceX;traceY;phi_b;phi_r;JRC;JSC;asperity;simgaC;cohesion;tension;
-0.0000000e+00;0;1.0000000e+00;5;1;7000;0;0;0;0;0;0;0;0;
-5.0000000e-01;0;8.6602540e-01;5;1;7000;0;0;0;0;0;0;0;0;
-8.6602540e-01;0;5.0000000e-01;5;1;7000;0;0;0;0;0;0;0;0;
-1.0000000e+00;0;6.1232340e-17;5;1;7000;0;0;0;0;0;0;0;0;
-8.6602540e-01;0;-5.0000000e-01;5;1;7000;0;0;0;0;0;0;0;0;
-5.0000000e-01;0;-8.6602540e-01;5;1;7000;0;0;0;0;0;0;0;0;
-5.6655389e-16;0;-1.0000000e+00;5;1;7000;0;0;0;0;0;0;0;0;
--5.0000000e-01;0;-8.6602540e-01;5;1;7000;0;0;0;0;0;0;0;0;
--8.6602540e-01;0;-5.0000000e-01;5;1;7000;0;0;0;0;0;0;0;0;
--1.0000000e+00;0;-1.8369702e-16;5;1.2;7000;0;0;0;0;0;0;0;0;
--8.6602540e-01;0;5.0000000e-01;5;1;7000;0;0;0;0;0;0;0;0;
--5.0000000e-01;0;8.6602540e-01;5;1;7000;0;0;0;0;0;0;0;0;
