--- examples/PotentialBlocks/Vaiont/2D/jointsSec2spacing/boundariesMatch.csv
@@ -1,4 +1,0 @@
-dip;dipdir;centreX;centreY;centreZ;phi_b;phi_r;JRC;JSC;asperity;simgaC;cohesion;tension;lambda0;heatCapacities;hwater;
-0;0;0;0;0;40;40;0;0;0;0;0;0.0;0;0.0;0.0;
-13.5;0;866;0;0;40;40;0;0;0;0;0;0.0;0;0.0;0.0;
-37.5;0;0;0;-575;40;40;0;0;0;0;0;0.0;0;0.0;0.0;
