--- examples/PotentialBlocks/Vaiont/2D/jointsSec2spacing/jointProbabilisticMatch40.csv
@@ -1,45 +1,0 @@
-dip;dipDir;centreX;centreY;centreZ;traceX;traceY;phi_b;phi_r;JRC;JSC;asperity;simgaC;cohesion;tension;
-90;0;352;0;-440;2000.0;2000.0;40;40;0;0;0;0;0.787e+06;0.076e+06;0; vertical dividor for upper crest;
-90;0;690;0;-230;2000.0;2000.0;40;40;0;0;0;0;0.787e+06;0.076e+06;0; vertical dividor for horizontal part;
-90;0;920;0;0;2000.0;2000.0;40;40;0;0;0;0;0.787e+06;0.076e+06;0; vertical dividor for slope face;
-38;0;920;0;-230;1000;1000;40;40;0;0;0;0;0.787e+06;0.076e+06;0; slope face;
-0;0;920;0;-230;10;10;40;40;0;0;0;0;0.787e+06;0.076e+06;0; horizontal part;
-33;0;352;0;-445;10;10;40;40;0;0;0;0;0.787e+06;0.076e+06;0; upper slope face;
-19;0;352;0;-445;100;100;40;40;0;0;0;0;0.787e+06;0.076e+06;0; upper crest face;
-0;0;1100;0;-30.0;350.0;350.0;40;40;0;0;0;0;0.787e+06;0.076e+06;0;;
-0;0;1100;0;-70.0;350.0;350.0;40;40;0;0;0;0;0.787e+06;0.076e+06;0;;
-0;0;1100;0;-110.0;350.0;350.0;40;40;0;0;0;0;0.787e+06;0.076e+06;0;;
-0;0;1100;0;-150.0;350.0;350.0;40;40;0;0;0;0;0.787e+06;0.076e+06;0;;
-0;0;1100;0;-190.0;350.0;350.0;40;40;0;0;0;0;0.787e+06;0.076e+06;0;;
-37.5;0;215;0;-560.0;400.0;400.0;40;40;0;0;0;0;0.787e+06;0.076e+06;0;;
-37.5;0;215;0;-520.0;400.0;400.0;40;40;0;0;0;0;0.787e+06;0.076e+06;0;;
-37.5;0;215;0;-480.0;200.0;200.0;40;40;0;0;0;0;0.787e+06;0.076e+06;0;;
-37.5;0;215;0;-440.0;200.0;200.0;40;40;0;0;0;0;0.787e+06;0.076e+06;0;;
-37.5;0;215;0;-400.0;200.0;200.0;40;40;0;0;0;0;0.787e+06;0.076e+06;0;;
-90;0;30;0;0;2000.0;2000.0;40;40;0;0;0;0;0.787e+06;0.076e+06;0;;
-90;0;70;0;0;2000.0;2000.0;40;40;0;0;0;0;0.787e+06;0.076e+06;0;;
-90;0;110;0;0;2000.0;2000.0;40;40;0;0;0;0;0.787e+06;0.076e+06;0;;
-90;0;150;0;0;2000.0;2000.0;40;40;0;0;0;0;0.787e+06;0.076e+06;0;;
-90;0;190;0;0;2000.0;2000.0;40;40;0;0;0;0;0.787e+06;0.076e+06;0;;
-90;0;230;0;0;2000.0;2000.0;40;40;0;0;0;0;0.787e+06;0.076e+06;0;;
-90;0;270;0;0;2000.0;2000.0;40;40;0;0;0;0;0.787e+06;0.076e+06;0;;
-90;0;310;0;0;2000.0;2000.0;40;40;0;0;0;0;0.787e+06;0.076e+06;0;;
-90;0;380;0;0;2000.0;2000.0;40;40;0;0;0;0;0.787e+06;0.076e+06;0;;
-90;0;410;0;0;2000.0;2000.0;40;40;0;0;0;0;0.787e+06;0.076e+06;0;;
-90;0;450;0;0;2000.0;2000.0;40;40;0;0;0;0;0.787e+06;0.076e+06;0;;
-90;0;490;0;0;2000.0;2000.0;40;40;0;0;0;0;0.787e+06;0.076e+06;0;;
-90;0;530;0;0;2000.0;2000.0;40;40;0;0;0;0;0.787e+06;0.076e+06;0;;
-90;0;570;0;0;2000.0;2000.0;40;40;0;0;0;0;0.787e+06;0.076e+06;0;;
-90;0;610;0;0;2000.0;2000.0;40;40;0;0;0;0;0.787e+06;0.076e+06;0;;
-90;0;650;0;0;2000.0;2000.0;40;40;0;0;0;0;0.787e+06;0.076e+06;0;;
-90;0;730;0;0;2000.0;2000.0;40;40;0;0;0;0;0.787e+06;0.076e+06;0;;
-90;0;770;0;0;2000.0;2000.0;40;40;0;0;0;0;0.787e+06;0.076e+06;0;;
-90;0;810;0;0;2000.0;2000.0;40;40;0;0;0;0;0.787e+06;0.076e+06;0;;
-90;0;890;0;0;2000.0;2000.0;40;40;0;0;0;0;0.787e+06;0.076e+06;0;;
-90;0;960;0;0;2000.0;2000.0;40;40;0;0;0;0;0.787e+06;0.076e+06;0;;
-90;0;1000;0;0;2000.0;2000.0;40;40;0;0;0;0;0.787e+06;0.076e+06;0;;
-90;0;1040;0;0;2000.0;2000.0;40;40;0;0;0;0;0.787e+06;0.076e+06;0;;
-90;0;1080;0;0;2000.0;2000.0;40;40;0;0;0;0;0.787e+06;0.076e+06;0;;
-90;0;1120;0;0;2000.0;2000.0;40;40;0;0;0;0;0.787e+06;0.076e+06;0;;
-90;0;1160;0;0;2000.0;2000.0;40;40;0;0;0;0;0.787e+06;0.076e+06;0;;
-90;0;850;0;0;2000.0;2000.0;40;40;0;0;0;0;0.787e+06;0.076e+06;0;joint at lower kink;
