--- examples/PotentialBlocks/Vaiont/3D/joints/boundaries.csv
@@ -1,7 +1,0 @@
-dip;dipdir;centreX;centreY;centreZ;phi_b;phi_r;JRC;JSC;asperity;simgaC;cohesion;tension;lambda0;heatCapacities;hwater;
-19.5;80;1865;750;-800;40;40;0;0;0;0;0;0;0;0.0;0.0;
-10.0;80;1865;1241;-620;40;40;0;0;0;0;0;0;0;0.0;0.0;
-43;353;827;2172;-1250;40;40;0;0;0;0;0;0;0;0.0;0.0;
-38;9.5;1086;879;-1100;40;40;0;0;0;0;0;0;0;0.0;0.0;
-20;50;1640;1211;-665;40;40;0;0;0;0;0;0;0;0.0;0.0;
-50;172.0;2345;1667;-590;40.0;40.0;0;0;0;0;0.0;0.0;0;0.0;0.0;
