--- examples/PotentialBlocks/Vaiont/3D/joints/jointProbabilistic40basic80Fast.csv
@@ -1,76 +1,0 @@
-dip;dipDir;centreX;centreY;centreZ;traceX;traceY;phi_b;phi_r;JRC;JSC;asperity;simgaC;cohesion;tension; jointType
-90;90;831;1350;-800;10000;10000;40;40;0;0;0;0;0.787e+06;0.076e+06;1; 		SLOPE PROFILE West vertical dividor;
-90;90;809;2172;-800;10000;10000;40;40;0;0;0;0;0.787e+06;0.076e+06;1; 		SLOPE PROFILE East vertical dividor  1980;
-40;8.0;1769;1622;-830;10;10;40;40;0;0;0;0;0.787e+06;0.076e+06;1; 			SLOPE PROFILE Mid bottom slope face;
-55;5.0;2040;2400;-500;5;5;40;40;0;0;0;0;0.787e+06;0.076e+06;1; 		SLOPE PROFILE East bottom slope face 2287;
-38;0.0;1769;1163;-860;10;10;40;40;0;0;0;0;0.787e+06;0.076e+06;1; 			SLOPE PROFILE West bottom slope face;
-90;0.0;1549;1163;-860;1000;10;40;40;0;0;0;0;0.787e+06;0.076e+06;1; 		SLOPE PROFILE West horizontal part vertical dividor;
-90;0.0;1549;1622;-860;1000;10;40;40;0;0;0;0;0.787e+06;0.076e+06;1; 		SLOPE PROFILE Mid horizontal part vertical dividor;
-90;0.0;1549;2350;-900;1000;10;40;40;0;0;0;0;0.787e+06;0.076e+06;1; 		SLOPE PROFILE East horizontal part vertical dividor;
-0;0.0;1649;1163;-860;10;10;40;40;0;0;0;0;0.787e+06;0.076e+06;1; 			SLOPE PROFILE West horizontal part ;
-0;0.0;1649;1622;-830;5;5;40;40;0;0;0;0;0.787e+06;0.076e+06;1; 			SLOPE PROFILE Mid horizontal part ;
-90;0.0;1769;1163;-860;10000;10000;40;40;0;0;0;0;0.787e+06;0.076e+06;1; 		SLOPE PROFILE Bottom slope face vertical dividor;
-90;90;1054;1750;-800;10000;60;40;40;0;0;0;0;0.787e+06;0.076e+06;1;  		SLOPE PROFILE Mid E vertical dividor - only divide upper part;
-90;90;1068;1590;-800;10000;60;40;40;0;0;0;0;0.787e+06;0.076e+06;1; 		SLOPE PROFILE Mid W vertical dividor - only divide upper part;
-22;0.0;1529;1632;-830;5;5;40;40;0;0;0;0;0.787e+06;0.076e+06;1; 			SLOPE PROFILE Mid mid upper slope;
-28;340.0;1529;1822;-800;5;5;40;40;0;0;0;0;0.787e+06;0.076e+06;1; 			SLOPE PROFILE Mid E upper slope;
-27;20.0;1529;1452;-850;5;5;40;40;0;0;0;0;0.787e+06;0.076e+06;1; 			SLOPE PROFILE Mid W upper slope;
-90;0.0;1225;1099;-860;10000;10000;40;40;0;0;0;0;0.787e+06;0.076e+06;1; 		SLOPE PROFILE West crest part vertical dividor;
-23;0.0;1215;1089;-1080;5;5;40;40;0;0;0;0;0.787e+06;0.076e+06;1; 			SLOPE PROFILE West crest;
-33;0.0;1540;1163;-870;5;5;40;40;0;0;0;0;0.787e+06;0.076e+06;1; 			SLOPE PROFILE West upper slope;
-30.0;0.0;689;2689;-1380;5000;5;40;40;0;0;0;0;0.787e+06;0.076e+06;1; 		SLOPE PROFILE East upper slope;
-90;90.0;1683;2690;-1250;5000;5000;36;36;0;0;0;0;0.0;0.0;1; 			PARRALLEL East vertical fault	CHECK THIS!!!;
-90;290.0;1683;2920;-1250;5000;5000;36;36;0;0;0;0;0.0;0.0;1; 			PARRALLEL East inclined fault	CHECK THIS!!!;
-37.5;0.0;1250;1500;-1140;150;450;40;40;0;0;0;0;0.787e+06;0.076e+06;1;		Joint parallel to slope (top) 180;
-37.5;0.0;1250;1500;-1060;150;450;40;40;0;0;0;0;0.787e+06;0.076e+06;1;		Joint parallel to slope (top) 180;
-37.5;0.0;1250;1500;-980;150;450;40;40;0;0;0;0;0.787e+06;0.076e+06;1; 		Joint parallel to slope (btm) 180;
-37.5;0.0;1250;1500;-900;150;450;40;40;0;0;0;0;0.787e+06;0.076e+06;1; 		Joint parallel to slope (mid lowest section);
-10.0;80.0;1800;1800;-570;100;550;40.00;40.00;0;0;0;0;0.787e+06;0.076e+06;1;			HORIZONTAL (btm) 9000x400;
-10.0;80.0;1800;1800;-650;100;550;40.00;40.00;0;0;0;0;0.787e+06;0.076e+06;1; 			HORIZONTAL (btm) 9000x400;
-10.0;80.0;1800;1800;-730;100;550;40.00;40.00;0;0;0;0;0.787e+06;0.076e+06;1; 	HORIZONTAL (btm) 9000x400;
-10.0;80.0;1800;1800;-810;250;650;40.00;40.00;0;0;0;0;0.787e+06;0.076e+06;1; 	HORIZONTAL (btm) new modified;
-10.0;80.0;1800;1800;-890;250;650;40.00;40.00;0;0;0;0;0.787e+06;0.076e+06;1; 	HORIZONTAL (btm) new modified;
-45;0.0;1250;2300;-880;150;100;40;40;0;0;0;0;0.787e+06;0.076e+06;1; 		Joint parallel to slope (btm) 180 modified;
-45;0.0;1250;2300;-960;150;100;40;40;0;0;0;0;0.787e+06;0.076e+06;1; 		Joint parallel to slope (btm) 180 modified;
-45;0.0;1250;2300;-1040;150;100;40;40;0;0;0;0;0.787e+06;0.076e+06;1; 		Joint parallel to slope (btm) 180 new;
-45;0.0;1250;2300;-1120;400;100;40;40;0;0;0;0;0.787e+06;0.076e+06;1; 		Joint parallel to slope (btm) 180 new;
-45;0.0;1250;2300;-1200;400;100;40;40;0;0;0;0;0.787e+06;0.076e+06;1; 		Joint parallel to slope (btm) 180 new;
-34;10;689;3589;-1350;5000;5000;40;40;0;0;0;0;0.787e+06;0.076e+06;1; 		Eastshave upper slope;
-90;90;831;1270;-800;10000;10000;40;40;0;0;0;0;0.787e+06;0.076e+06;1; 		SLOPE PROFILE parallel;
-90;90;831;1190;-800;10000;10000;40;40;0;0;0;0;0.787e+06;0.076e+06;1; 		SLOPE PROFILE parallel;
-90;90;831;1110;-800;10000;10000;40;40;0;0;0;0;0.787e+06;0.076e+06;1; 		SLOPE PROFILE parallel;
-90;90;831;1030;-800;10000;10000;40;40;0;0;0;0;0.787e+06;0.076e+06;1; 		SLOPE PROFILE parallel;
-90;90;831;950;-800;10000;10000;40;40;0;0;0;0;0.787e+06;0.076e+06;1; 		SLOPE PROFILE parallel;
-90;90;831;870;-800;10000;10000;40;40;0;0;0;0;0.787e+06;0.076e+06;1; 		SLOPE PROFILE parallel;
-90;90;831;790;-800;10000;10000;40;40;0;0;0;0;0.787e+06;0.076e+06;1; 		SLOPE PROFILE parallel;
-90;90;831;710;-800;10000;10000;40;40;0;0;0;0;0.787e+06;0.076e+06;1; 		SLOPE PROFILE parallel;
-90;90;831;630;-800;10000;10000;40;40;0;0;0;0;0.787e+06;0.076e+06;1; 		SLOPE PROFILE parallel;
-90;90;831;1350;-800;10000;10000;40;40;0;0;0;0;0.787e+06;0.076e+06;1; 		SLOPE PROFILE parallel;
-90;90;831;1430;-800;10000;10000;40;40;0;0;0;0;0.787e+06;0.076e+06;1; 		SLOPE PROFILE parallel;
-90;90;831;1510;-800;10000;10000;40;40;0;0;0;0;0.787e+06;0.076e+06;1; 		SLOPE PROFILE parallel;
-90;90;831;1590;-800;10000;10000;40;40;0;0;0;0;0.787e+06;0.076e+06;1; 		SLOPE PROFILE parallel;
-90;90;831;1670;-800;10000;10000;40;40;0;0;0;0;0.787e+06;0.076e+06;1; 		SLOPE PROFILE parallel;
-90;90;831;1750;-800;10000;10000;40;40;0;0;0;0;0.787e+06;0.076e+06;1; 		SLOPE PROFILE parallel;
-90;90;831;1830;-800;10000;10000;40;40;0;0;0;0;0.787e+06;0.076e+06;1; 		SLOPE PROFILE parallel;
-90;90;831;1910;-800;10000;10000;40;40;0;0;0;0;0.787e+06;0.076e+06;1; 		SLOPE PROFILE parallel;
-90;90;831;1990;-800;10000;10000;40;40;0;0;0;0;0.787e+06;0.076e+06;1; 		SLOPE PROFILE parallel;
-90;90;831;2080;-800;10000;10000;40;40;0;0;0;0;0.787e+06;0.076e+06;1; 		SLOPE PROFILE parallel;
-90;90;831;2250;-800;10000;10000;40;40;0;0;0;0;0.787e+06;0.076e+06;1; 		SLOPE PROFILE parallel;
-90;90;831;2330;-800;10000;10000;40;40;0;0;0;0;0.787e+06;0.076e+06;1; 		SLOPE PROFILE parallel;
-90;90;831;2410;-800;10000;10000;40;40;0;0;0;0;0.787e+06;0.076e+06;1; 		SLOPE PROFILE parallel;
-90;90;831;2490;-800;10000;10000;40;40;0;0;0;0;0.787e+06;0.076e+06;1; 		SLOPE PROFILE parallel;
-90;90;831;2560;-800;10000;10000;40;40;0;0;0;0;0.787e+06;0.076e+06;1; 		SLOPE PROFILE parallel;
-90;90;831;2630;-800;10000;10000;40;40;0;0;0;0;0.787e+06;0.076e+06;1; 		SLOPE PROFILE parallel;
-90;0.0;1850;1163;-860;10000;1500;40;40;0;0;0;0;0.787e+06;0.076e+06;1; 		SLOPE PROFILE perpendicular bottom 1769;
-90;0.0;1930;1163;-860;10000;1500;40;40;0;0;0;0;0.787e+06;0.076e+06;1; 		SLOPE PROFILE perpendicular;
-90;0.0;2010;1163;-860;10000;1500;40;40;0;0;0;0;0.787e+06;0.076e+06;1; 		SLOPE PROFILE perpendicular;
-90;0.0;2090;1163;-860;10000;1500;40;40;0;0;0;0;0.787e+06;0.076e+06;1; 		SLOPE PROFILE perpendicular;
-90;0.0;1690;1163;-860;10000;1500;40;40;0;0;0;0;0.787e+06;0.076e+06;1; 		SLOPE PROFILE perpendicular;
-90;0.0;1610;1163;-860;10000;1500;40;40;0;0;0;0;0.787e+06;0.076e+06;1; 		SLOPE PROFILE perpendicular mid 1549;
-90;0.0;1470;1163;-860;10000;1500;40;40;0;0;0;0;0.787e+06;0.076e+06;1; 		SLOPE PROFILE perpendicular ;
-90;0.0;1390;1163;-860;10000;1500;40;40;0;0;0;0;0.787e+06;0.076e+06;1; 		SLOPE PROFILE perpendicular ;
-90;0.0;1310;1163;-860;10000;1500;40;40;0;0;0;0;0.787e+06;0.076e+06;1; 		SLOPE PROFILE perpendicular upper 1225;
-90;0.0;1140;1163;-860;10000;1500;40;40;0;0;0;0;0.787e+06;0.076e+06;1; 		SLOPE PROFILE perpendicular;
-90;0.0;1060;1163;-860;10000;1500;40;40;0;0;0;0;0.787e+06;0.076e+06;1; 		SLOPE PROFILE perpendicular;
-90;0.0;980;1163;-860;10000;1500;40;40;0;0;0;0;0.787e+06;0.076e+06;1; 		SLOPE PROFILE perpendicular;
-90;0.0;900;1163;-860;10000;1500;40;40;0;0;0;0;0.787e+06;0.076e+06;1; 		SLOPE PROFILE perpendicular;
