--- examples/PotentialBlocks/Vaiont/3D/joints/sliceBoundaries.csv
@@ -1,4 +1,0 @@
-dip;dipdir;centreX;centreY;centreZ;phi_b;phi_r;JRC;JSC;asperity;simgaC;cohesion;tension
-85;0.0;2130;1653;-590;40;40;0;0;0;0;0;0;
-75;176.0;2230;1667;-590;40;40;0;0;0;0;0;0;
-90;90.0;1683;2690;-1250;40;40;0;0;0;0;0;0;
