--- examples/PotentialBlocks/Vaiont/3D/saveData/fullJoint40degWetBasic80softeningKs120161020T231226p3349/Vaiont3Ddata.txt
@@ -1,1 +1,0 @@
-timeStep	time	unbalancedForce	kineticEnergy	waterLevel	boundary_phi	displacementW	displacementM	displacementE	displacementWx	displacementMx	displacementEx	displacementWy	displacementMy	displacementEy	velocityW	velocityM	velocityE	CohesionBrokenNo	TensionBrokenNo
