+++ examples/PotentialBlocks/joints/jointC.csv
@@ -1,0 +1,74 @@
+dip;dipDir;centreX;centreY;centreZ;traceX;traceY;phi_b;phi_r;JRC;JSC;asperity;simgaC;cohesion;tension; jointType;
+90;315;0;0;0;10000;10000;40;40;0;0;0;0;787000;76000;1;joint;
+70;160;-52.8;19.2;-20.4;10000;10000;40;40;0;0;0;0;787000;76000;1;joint;
+70;160;-44;16;-17;10000;10000;40;40;0;0;0;0;787000;76000;1;joint;
+70;160;-35.2;12.8;-13.6;10000;10000;40;40;0;0;0;0;787000;76000;1;joint;
+70;160;-26.4;9.6;-10.2;10000;10000;40;40;0;0;0;0;787000;76000;1;joint;
+70;160;-17.6;6.4;-6.8;10000;10000;40;40;0;0;0;0;787000;76000;1;joint;
+70;160;-8.8;3.2;-3.4;10000;10000;40;40;0;0;0;0;787000;76000;1;joint;
+70;160;0;0;0;10000;10000;40;40;0;0;0;0;787000;76000;1;joint;
+70;160;8.8;-3.2;3.4;10000;10000;40;40;0;0;0;0;787000;76000;1;joint;
+70;160;17.6;-6.4;6.8;10000;10000;40;40;0;0;0;0;787000;76000;1;joint;
+70;160;26.4;-9.6;10.2;10000;10000;40;40;0;0;0;0;787000;76000;1;joint;
+70;160;35.2;-12.8;13.6;10000;10000;40;40;0;0;0;0;787000;76000;1;joint;
+70;160;44;-16;17;10000;10000;40;40;0;0;0;0;787000;76000;1;joint;
+70;160;52.8;-19.2;20.4;10000;10000;40;40;0;0;0;0;787000;76000;1;joint;
+70;160;61.6;-22.4;23.8;10000;10000;40;40;0;0;0;0;787000;76000;1;joint;
+70;160;70.4;-25.6;27.2;10000;10000;40;40;0;0;0;0;787000;76000;1;joint;
+70;160;79.2;-28.8;30.6;10000;10000;40;40;0;0;0;0;787000;76000;1;joint;
+60;90;0;68.8;-40;10000;10000;40;40;0;0;0;0;787000;76000;1;joint;
+60;90;0;60.2;-35;10000;10000;40;40;0;0;0;0;787000;76000;1;joint;
+60;90;0;51.6;-30;10000;10000;40;40;0;0;0;0;787000;76000;1;joint;
+60;90;0;43;-25;10000;10000;40;40;0;0;0;0;787000;76000;1;joint;
+60;90;0;34.4;-20;10000;10000;40;40;0;0;0;0;787000;76000;1;joint;
+60;90;0;25.8;-15;10000;10000;40;40;0;0;0;0;787000;76000;1;joint;
+60;90;0;17.2;-10;10000;10000;40;40;0;0;0;0;787000;76000;1;joint;
+60;90;0;8.6;-5;10000;10000;40;40;0;0;0;0;787000;76000;1;joint;
+60;90;0;0;0;10000;10000;40;40;0;0;0;0;787000;76000;1;joint;
+60;90;0;-8.6;5;10000;10000;40;40;0;0;0;0;787000;76000;1;joint;
+60;90;0;-17.2;10;10000;10000;40;40;0;0;0;0;787000;76000;1;joint;
+60;90;0;-25.8;15;10000;10000;40;40;0;0;0;0;787000;76000;1;joint;
+60;90;0;-34.4;20;10000;10000;40;40;0;0;0;0;787000;76000;1;joint;
+60;90;0;-43;25;10000;10000;40;40;0;0;0;0;787000;76000;1;joint;
+60;90;0;-51.6;30;10000;10000;40;40;0;0;0;0;787000;76000;1;joint;
+60;90;0;-60.2;35;10000;10000;40;40;0;0;0;0;787000;76000;1;joint;
+60;90;0;-68.8;40;10000;10000;40;40;0;0;0;0;787000;76000;1;joint;
+60;90;0;-77.4;45;10000;10000;40;40;0;0;0;0;787000;76000;1;joint;
+60;90;0;-86;50;10000;10000;40;40;0;0;0;0;787000;76000;1;joint;
+70;350;82.8;-14.4;-30.6;10000;10000;40;40;0;0;0;0;787000;76000;1;joint;
+70;350;73.6;-12.8;-27.2;10000;10000;40;40;0;0;0;0;787000;76000;1;joint;
+70;350;64.4;-11.2;-23.8;10000;10000;40;40;0;0;0;0;787000;76000;1;joint;
+70;350;55.2;-9.6;-20.4;10000;10000;40;40;0;0;0;0;787000;76000;1;joint;
+70;350;46;-8;-17;10000;10000;40;40;0;0;0;0;787000;76000;1;joint;
+70;350;36.8;-6.4;-13.6;10000;10000;40;40;0;0;0;0;787000;76000;1;joint;
+70;350;27.6;-4.8;-10.2;10000;10000;40;40;0;0;0;0;787000;76000;1;joint;
+70;350;18.4;-3.2;-6.8;10000;10000;40;40;0;0;0;0;787000;76000;1;joint;
+70;350;9.2;-1.6;-3.4;10000;10000;40;40;0;0;0;0;787000;76000;1;joint;
+70;350;-3.5527136788005E-15;-8.88178419700125E-16;0;10000;10000;40;40;0;0;0;0;787000;76000;1;joint;
+70;350;-9.2;1.6;3.4;10000;10000;40;40;0;0;0;0;787000;76000;1;joint;
+70;350;-18.4;3.2;6.8;10000;10000;40;40;0;0;0;0;787000;76000;1;joint;
+70;350;-27.6;4.8;10.2;10000;10000;40;40;0;0;0;0;787000;76000;1;joint;
+70;350;-36.8;6.4;13.6;10000;10000;40;40;0;0;0;0;787000;76000;1;joint;
+70;350;-46;8;17;10000;10000;40;40;0;0;0;0;787000;76000;1;joint;
+70;350;-55.2;9.6;20.4;10000;10000;40;40;0;0;0;0;787000;76000;1;joint;
+70;350;-64.4;11.2;23.8;10000;10000;40;40;0;0;0;0;787000;76000;1;joint;
+70;350;-73.6;12.8;27.2;10000;10000;40;40;0;0;0;0;787000;76000;1;joint;
+70;350;-82.8;14.4;30.6;10000;10000;40;40;0;0;0;0;787000;76000;1;joint;
+60;240;-30.1;-52.5;-35;10000;10000;40;40;0;0;0;0;787000;76000;1;joint;
+60;240;-25.8;-45;-30;10000;10000;40;40;0;0;0;0;787000;76000;1;joint;
+60;240;-21.5;-37.5;-25;10000;10000;40;40;0;0;0;0;787000;76000;1;joint;
+60;240;-17.2;-30;-20;10000;10000;40;40;0;0;0;0;787000;76000;1;joint;
+60;240;-12.9;-22.5;-15;10000;10000;40;40;0;0;0;0;787000;76000;1;joint;
+60;240;-8.6;-15;-10;10000;10000;40;40;0;0;0;0;787000;76000;1;joint;
+60;240;-4.3;-7.5;-5;10000;10000;40;40;0;0;0;0;787000;76000;1;joint;
+60;240;3.5527136788005E-15;0;0;10000;10000;40;40;0;0;0;0;787000;76000;1;joint;
+60;240;4.3;7.5;5;10000;10000;40;40;0;0;0;0;787000;76000;1;joint;
+60;240;8.6;15;10;10000;10000;40;40;0;0;0;0;787000;76000;1;joint;
+60;240;12.9;22.5;15;10000;10000;40;40;0;0;0;0;787000;76000;1;joint;
+60;240;17.2;30;20;10000;10000;40;40;0;0;0;0;787000;76000;1;joint;
+60;240;21.5;37.5;25;10000;10000;40;40;0;0;0;0;787000;76000;1;joint;
+60;240;25.8;45;30;10000;10000;40;40;0;0;0;0;787000;76000;1;joint;
+60;240;30.1;52.5;35;10000;10000;40;40;0;0;0;0;787000;76000;1;joint;
+60;240;34.4;60;40;10000;10000;40;40;0;0;0;0;787000;76000;1;joint;
+60;240;38.7;67.5;45;10000;10000;40;40;0;0;0;0;787000;76000;1;joint;
+60;240;43;75;50;10000;10000;40;40;0;0;0;0;787000;76000;1;joint;
